--- onedal/common/device_lookup.hpp
@@ -1,31 +1,0 @@
-/*******************************************************************************
-* Copyright 2024 Intel Corporation
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*******************************************************************************/
-
-#include <cstdint>
-#include <optional>
-
-#include "oneapi/dal/detail/policy.hpp"
-
-namespace oneapi::dal::python {
-
-#ifdef ONEDAL_DATA_PARALLEL
-
-std::optional<sycl::device> get_device_by_id(std::uint32_t id);
-std::optional<std::uint32_t> get_device_id(const sycl::device& device);
-
-#endif // ONEDAL_DATA_PARALLEL
-
-} // namespace oneapi::dal::python
